cleaning program class
