+++ Talabat.APIs/Extentions/SwaggerServiceExtention.cs
@@ -1,0 +1,28 @@
+﻿using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Mvc;
+using Talabat.APIs.Helpers;
+using Talabat.Core.Repository.Contract;
+using Talabat.Repository;
+
+namespace Talabat.APIs.Extentions
+{
+    public static class SwaggerServiceExtention
+    {
+        public static IServiceCollection SwaggerServices(this IServiceCollection services)
+        {
+            // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
+            //about swagger
+            services.AddEndpointsApiExplorer();
+            services.AddSwaggerGen();
+
+            return services;
+        }
+        public static WebApplication SwaggerMiddleWares(this WebApplication app)
+        {
+            app.UseSwagger();
+            app.UseSwaggerUI();
+
+            return app;
+        }
+    }
+}
